refactor: extract WatchMixin from NexusFS (first mixin extraction)
Move sys_watch() to nexus_fs_watch.py as WatchMixin. NexusFS inherits
from WatchMixin (before ABC in MRO for abstract method resolution).

Pattern for subsequent extractions: mixin is plain class with
TYPE_CHECKING stubs for kernel primitives, listed before
NexusFilesystemABC in inheritance.
